+++ libs/core/bar/bar.component.ts
@@ -2,13 +2,14 @@ import {
     ChangeDetectionStrategy,
     Component,
     ElementRef,
+    input,
     Input,
     OnChanges,
     OnDestroy,
     OnInit,
     ViewEncapsulation
 } from '@angular/core';
-import { CssClassBuilder, applyCssClass } from '@fundamental-ngx/cdk/utils';
+import { applyCssClass, CssClassBuilder } from '@fundamental-ngx/cdk/utils';
 import {
     ContentDensityMode,
     ContentDensityObserver,
@@ -36,6 +37,9 @@ export type BarDesignType = 'header' | 'subheader' | 'header-with-subheader' | '
             defaultContentDensity: ContentDensityMode.COMPACT
         })
     ],
+    host: {
+        '[attr.role]': 'role()'
+    },
     standalone: true
 })
 export class BarComponent implements OnChanges, OnInit, CssClassBuilder, OnDestroy {
@@ -62,6 +66,11 @@ export class BarComponent implements OnChanges, OnInit, CssClassBuilder, OnDestr
      */
     @Input()
     size: SizeType = '';
+
+    /** Aria role for the Bar
+     * default is toolbar
+     */
+    role = input('toolbar');
 
     /** @hidden */
     private _subscriptions = new Subscription();