+++ libs/core/message-box/message-box-semantic-icon/message-box-semantic-icon.component.ts
@@ -13,7 +13,11 @@ import { MessageBoxConfig, MessageBoxHost } from '../utils/message-box-config.cl
     selector: 'fd-message-box-semantic-icon',
     template: `<fd-icon [glyph]="_getIcon" [font]="glyphFont" role="presentation"></fd-icon>`,
     changeDetection: ChangeDetectionStrategy.OnPush,
-    imports: [IconComponent]
+    imports: [IconComponent],
+    host: {
+        role: 'presentation',
+        '[attr.aria-hidden]': 'true'
+    }
 })
 export class MessageBoxSemanticIconComponent {
     /** Custom semantic icon */