+++ libs/docs/core/dialog/examples/dialog-backdrop-container/dialog-backdrop-container-example.component.ts
@@ -50,8 +50,8 @@ export class DialogBackdropContainerExampleComponent {
             responsivePadding: true,
             backdropClass: 'dialog-custom-overlay-example',
             data: `This dialog has a custom backdrop!`,
-            ariaLabelledBy: 'fd-dialog-header-2',
-            ariaDescribedBy: 'fd-dialog-body-2'
+            ariaLabelledBy: 'fd-dialog-header-backdrop',
+            ariaDescribedBy: 'fd-dialog-body-backdrop'
         });
     }
 
@@ -61,8 +61,8 @@ export class DialogBackdropContainerExampleComponent {
             container: containerRef,
             responsivePadding: true,
             data: `This dialog has been opened inside a local div!`,
-            ariaLabelledBy: 'fd-dialog-header-2',
-            ariaDescribedBy: 'fd-dialog-body-2'
+            ariaLabelledBy: 'fd-dialog-header-backdrop',
+            ariaDescribedBy: 'fd-dialog-body-backdrop'
         });
     }
 
@@ -74,8 +74,8 @@ export class DialogBackdropContainerExampleComponent {
             responsivePadding: true,
             backdropClass: 'static-dialog',
             data: `This dialog has been opened inside a local div and displayed as a static element!`,
-            ariaLabelledBy: 'fd-dialog-header-2',
-            ariaDescribedBy: 'fd-dialog-body-2'
+            ariaLabelledBy: 'fd-dialog-header-backdrop',
+            ariaDescribedBy: 'fd-dialog-body-backdrop'
         });
     }
 }